+++ scout/adapter/mongo/filter.py
@@ -4,6 +4,7 @@ import logging
 import pymongo
 from bson.objectid import ObjectId
 from flask import url_for
+from werkzeug.datastructures import MultiDict
 
 LOG = logging.getLogger(__name__)
 
@@ -37,16 +38,18 @@ class FilterHandler(object):
         return filter_obj
 
     def stash_filter(
-        self, filter_obj, institute_obj, case_obj, user_obj, category="snv", link=None
-    ):
+        self,
+        filter_obj: MultiDict,
+        institute_obj: dict,
+        case_obj: dict,
+        user_obj: dict,
+        category: str = "snv",
+        link: str = None,
+    ) -> str:
         """Store away filter settings for later use.
 
         Arguments:
-            filter_obj(MultiDict)
-            institute_obj(dict)
-            user_obj(dict)
-            case_obj(dict)
-            category(str): in ['cancer', 'snv', 'str', 'sv']
+            category(str): in ['cancer', 'snv', 'str', 'sv', ...]
 
         Returns:
             filter_id(str) - a unique id that can be cast to ObjectId
@@ -228,15 +231,12 @@ class FilterHandler(object):
 
         return result
 
-    def filters(self, institute_id, category="snv"):
+    def filters(self, institute_id: str, category: str = "snv") -> pymongo.cursor.Cursor:
         """Obtain a cursor for all filters available to an institute in a category.
 
         Arguments:
-            institute_id(str)
-            category(str): in ['cancer', 'snv', 'str', 'sv']
+            category(str): in ['cancer', 'snv', 'str', 'sv', ...]
 
-        Returns:
-            filters(pymongo.Cursor)
         """
         filters_res = self.filter_collection.find(
             {"institute_id": institute_id, "category": category}